requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
rq==1.16.2
redis==5.0.4
qdrant-client==1.12.1
sqlalchemy==2.0.35
python-dotenv==1.0.1
trafilatura==2.0.0
google-genai==0.2.2
httpx==0.27.0
tqdm==4.66.5
